+++ aihub_agent/aihub_agent/agents/RetrievalAgent/configs/RetrievalAgentConfig.py
@@ -1,7 +1,7 @@
 from typing import Annotated
 
 from aihub_lib.agents.AgentConfig import AgentConfig
-from aihub_lib.generative_ai.retrievers import KnowledgeRetrieverConfig
+from aihub_lib.generative_ai.retrievers.KnowledgeRetrieverConfig import KnowledgeRetrieverConfig
 from aihub_lib.i18n.LocaleString import LocaleString
 from pydantic import Field
