+++ aihub_lib/aihub_lib/generative_ai/retrievers/create_retriever.py
@@ -1,11 +1,9 @@
-from aihub_lib.generative_ai.retrievers import (
-    BaseRetriever,
-    BaseRetrieverConfig,
-    InsightRetriever,
-    InsightRetrieverConfig,
-    KnowledgeRetriever,
-    KnowledgeRetrieverConfig,
-)
+from aihub_lib.generative_ai.retrievers.BaseRetriever import BaseRetriever
+from aihub_lib.generative_ai.retrievers.BaseRetrieverConfig import BaseRetrieverConfig
+from aihub_lib.generative_ai.retrievers.InsightRetriever import InsightRetriever
+from aihub_lib.generative_ai.retrievers.InsightRetrieverConfig import InsightRetrieverConfig
+from aihub_lib.generative_ai.retrievers.KnowledgeRetriever import KnowledgeRetriever
+from aihub_lib.generative_ai.retrievers.KnowledgeRetrieverConfig import KnowledgeRetrieverConfig
 
 
 def create_retriever(config: BaseRetrieverConfig) -> BaseRetriever:
